split reducers file into index.js and recipeReducer.js, renamed one action type
