+++ client/src/actions/index.js
@@ -3,6 +3,8 @@ import axios from 'axios';
 export const ERROR = 'ERROR';
 export const GET_RECIPES = 'GET_RECIPES';
 export const GETTING_RECIPES = 'GETTING_RECIPES';
+export const RECIPE_SUCCESS = 'RECIPE_SUCCESS';
+export const RECIPE_FAILURE = 'RECIPE_FAILURE';
 
 const URL = "https://donteatthat.herokuapp.com/"
 
@@ -22,18 +24,17 @@ export const getRecipes = () => {
   };
 };
 
-
 // single recipe
-// export const getRecipe = () => {
-//     const recipes = axios.get(`${URL}api/recipes/all`);
-//     return dispatch => {
-//       dispatch({ type: GETTING_RECIPE });
-//       recipes
-//         .then(response => {
-//           dispatch({ type: GET_RECIPES, payload: response.data });
-//         })
-//         .catch(err => {
-//           dispatch({ type: ERROR, payload: err });
-//         });
-//     };
-//   };
+export const getRecipe = id => {
+    const recipes = axios.get(`${URL}api/recipes/one/${id}`);
+    return dispatch => {
+      dispatch({ type: GETTING_RECIPES });
+      recipes
+        .then(response => {
+          dispatch({ type: GET_RECIPES, payload: response.data });
+        })
+        .catch(err => {
+          dispatch({ type: ERROR, payload: err });
+        });
+    };
+  };
